+++ data-migrations/2024-12-17-1133-create-project-tables.sql
@@ -22,6 +22,7 @@ CREATE TABLE `projectContributors` (
   `id` INT AUTO_INCREMENT PRIMARY KEY,
   `projectId` INT NOT NULL,
   `affiliationId` VARCHAR(255),
+  `userId` INT NOT NULL,
   `givenName` VARCHAR(255),
   `surName` VARCHAR(255),
   `orcid` VARCHAR(255),
@@ -45,6 +46,25 @@ CREATE TABLE `projectContributorRoles` (
   `modified` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
   FOREIGN KEY (projectContributorId) REFERENCES projectContributors(id) ON DELETE CASCADE,
   FOREIGN KEY (contributorRoleId) REFERENCES contributorRoles(id) ON DELETE CASCADE
+) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb3;
+
+# People who have access to collaborate on writing the Project
+CREATE TABLE `projectCollaborators` (
+  `id` INT AUTO_INCREMENT PRIMARY KEY,
+  `projectId` INT NOT NULL,
+  `email` VARCHAR(255) NOT NULL,
+  `invitedById` INT NOT NULL,
+  `userId` INT NOT NULL,
+  `created` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+  `createdById` INT NOT NULL,
+  `modified` TIMESTAMP NOT NULL DEFAULT CURRENT_TIMESTAMP,
+  `modifiedById` INT NOT NULL,
+  CONSTRAINT unique_project_collaborator UNIQUE (`projectId`, `email`),
+  FOREIGN KEY (projectId) REFERENCES projects(id) ON DELETE CASCADE,
+  FOREIGN KEY (invitedById) REFERENCES users(id),
+  FOREIGN KEY (userId) REFERENCES users(id) ON DELETE CASCADE,
+  INDEX planCollaborators_email_idx (`email`),
+  INDEX planCollaborators_user_idx (`userId`)
 ) ENGINE=InnoDB AUTO_INCREMENT=1 DEFAULT CHARSET=utf8mb3;
 
 # Institutions/organizations that provide financial support for a research project